Moved a few things
Temporarily removed potion effects, currently migrating to Listener.
